feat: preconnect to pi server before unlocking
